+++ NEWS.rst
@@ -3,6 +3,14 @@ Release notes
 
 This project uses `semantic versioning <http://semver.org/>`_.
 
+master
+------
+
+Fixed
+^^^^^
+
+- ``Just.fmap()`` doesn't return Nothing appropriately.
+
 0.2.0
 -----